Move plugin header file to the plugin directory
